requirements.txt
Appium-Python-Client==2.0.0
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
cryptography==41.0.7
exceptiongroup==1.2.0
future==0.18.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Logbook==1.7.0.post0
outcome==1.3.0.post0
packaging==23.2
pluggy==0.13.1
py==1.11.0
pycparser==2.21
pyOpenSSL==23.3.0
PySocks==1.7.1
pytest==6.1.1
pytest-logbook==1.2.0
pytest-logger==0.5.1
selenium==4.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.23.2
trio-websocket==0.11.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wsproto==1.2.0
